Surveys util module: splitted into several files
